Combine various logs into a single GitHub comment.
This takes a very naive approach to combining comments; basically we
treat each log as a separate section of the combined comment and use
HTML comments to delimit the start and end of each section. This is
simple to implement but has a huge disadvantage that we are almost
certain to run into the GitHub comment length limits much more
often. It's not clear what a good solution to this is. One way forward
would be to actually parse out the old data and add to the existing
tables rather than having one table per test run. But that seems
hard. Another approach is simply to put much less pass/fail data in
the logs by default.